moved reset password to me endpoint
