+++ src/routes/me.js
@@ -1,5 +1,6 @@
 const Feed = require('../models/feed');
 const User = require('../models/user');
+const jwt = require('../utils/jwt');
 const ERROR_MESSAGES = require('../utils/error-messages');
 
 /**
@@ -112,5 +113,62 @@ exports.deleteMe = function(req, res) {
           id: ERROR_MESSAGES.INVALID_ID
         }
       });
+    });
+};
+
+
+/**
+ * POST /me/update-password
+ * Updates password for a user
+ * @param {Request} req
+ * @param {Response} res
+ */
+exports.updatePassword = function(req, res) {
+
+  let {
+    old_password,
+    new_password,
+    reset_password_token } = req.body;
+
+  let { user } = req;
+
+  let verify;
+
+  if (reset_password_token) {
+    let user = jwt.decode(reset_password_token);
+    verify = User.findUser({id:user.id});
+  } else {
+    verify = user.comparePassword(old_password).then((isValidPassword) => {
+      if (!isValidPassword) {
+        throw User.ERROR.INVALID_PASSWORD;
+      }
+
+      return user;
+    });
+  }
+
+  verify
+    .then((user) => {
+      user.password = new_password;
+      return user.save();
+    })
+    .then((user) => {
+      res.status(200).json({
+        id: user._id,
+        email: user.email
+      });
+    })
+    .catch((err) => {
+      let errors = {};
+
+      if (err === User.ERROR.INVALID_PASSWORD) {
+        errors.old_password = ERROR_MESSAGES.INVALID_OLD_PASSWORD;
+      } else if (err === User.ERROR.USER_NOT_FOUND) {
+        errors.id = ERROR_MESSAGES.USER_NOT_EXISTS;
+      } else {
+        errors.new_password = ERROR_MESSAGES.INVALID_PASSWORD;
+      }
+
+      res.status(400).json({errors});
     });
 };
